Reorganise repositories by module ownership
Move repository interfaces and implementations from generic src/Repositories/
folder into their respective module directories. This makes module boundaries
explicit and enables better encapsulation of data access concerns per module.

New structure:
- PersonRepository moved to src/Modules/Person/
- OrganisationRepository moved to src/Modules/Organisation/
- LogBookRepository moved to src/Modules/LogBook/
- StationRepository moved to src/Modules/Station/
- SentryRepository remains in src/Modules/Sentry/ (was already there)

All namespaces updated from DigitalGuardBook.Repositories to
DigitalGuardBook.Modules.{ModuleName}. Type aliases introduced
(e.g., StationEntity, LogBookEntity, OrganisationEntity, PersonEntity)
to disambiguate entity types from module namespace names.

Updated all importing files:
- src/Program.cs DI registration
- src/GraphQL/DigitalGuardBookQuery.cs and resolvers
- src/Modules/Sentry/SentryService.cs
- src/Modules/Sentry/EventHandlers/SentryLogBookEventHandler.cs
- Test files in DigitalGuardBook.Tests/

The src/Repositories/ directory is now empty and removed.
