+++ DigitalGuardBook.Tests/Modules/Sentry/SentryLogBookEventHandlerTests.cs
@@ -3,7 +3,8 @@ using DigitalGuardBook.Data.Entities;
 using DigitalGuardBook.Infrastructure;
 using DigitalGuardBook.Modules.Sentry.EventHandlers;
 using DigitalGuardBook.Modules.Sentry.Events;
-using DigitalGuardBook.Repositories;
+using DigitalGuardBook.Modules.LogBook;
+using DigitalGuardBook.Modules.Person;
 using Microsoft.Extensions.Localization;
 using Moq;
 using Xunit;
