+++ DigitalGuardBook.Tests/Modules/Sentry/SentryServiceTests.cs
@@ -3,7 +3,8 @@ using DigitalGuardBook.Data.Entities;
 using DigitalGuardBook.Infrastructure;
 using DigitalGuardBook.Modules.Sentry;
 using DigitalGuardBook.Modules.Sentry.Events;
-using DigitalGuardBook.Repositories;
+using DigitalGuardBook.Modules.Person;
+using DigitalGuardBook.Modules.Organisation;
 using Moq;
 using Xunit;
 using SentryEntity = DigitalGuardBook.Data.Entities.Sentry;
