+++ src/Modules/LogBook/ILogBookRepository.cs
@@ -1,10 +1,11 @@
 using DigitalGuardBook.Data.Entities;
+using LogBookEntity = DigitalGuardBook.Data.Entities.LogBook;
 
-namespace DigitalGuardBook.Repositories;
+namespace DigitalGuardBook.Modules.LogBook;
 
 public interface ILogBookRepository
 {
-    Task<LogBook> GetLogBookAsync(string stationId, int year);
+    Task<LogBookEntity> GetLogBookAsync(string stationId, int year);
     Task<List<LogBookEntry>> GetEntriesForLogBookAsync(DateTimeOffset? from, DateTimeOffset? to);
     Task InsertLogBookEntryAsync(string message, DateTimeOffset? time);
 }
