+++ src/Modules/Station/IStationRepository.cs
@@ -1,0 +1,10 @@
+using DigitalGuardBook.Data.Entities;
+using StationEntity = DigitalGuardBook.Data.Entities.Station;
+
+namespace DigitalGuardBook.Modules.Station;
+
+public interface IStationRepository
+{
+    Task<IList<StationEntity>> AllStationsAsync();
+    Task<StationEntity> GetStationAsync();
+}
