+++ src/Modules/Station/StationRepository.cs
@@ -2,8 +2,9 @@ using MongoDB.Driver;
 using MongoDB.Driver.Linq;
 using DigitalGuardBook.Data;
 using DigitalGuardBook.Data.Entities;
+using StationEntity = DigitalGuardBook.Data.Entities.Station;
 
-namespace DigitalGuardBook.Repositories
+namespace DigitalGuardBook.Modules.Station
 {
     public class StationRepository : IStationRepository
     {
@@ -14,14 +15,14 @@ namespace DigitalGuardBook.Repositories
             _dataContext = dataContext;
         }
 
-        public async Task<IList<Station>> AllStationsAsync()
+        public async Task<IList<StationEntity>> AllStationsAsync()
         {
             return await _dataContext.Stations
                 .AsQueryable()
                 .ToListAsync();
         }
 
-        public async Task<Station> GetStationAsync()
+        public async Task<StationEntity> GetStationAsync()
         {
             return await _dataContext.Stations
                 .AsQueryable()
