--- src/Repositories/IStationRepository.cs
@@ -1,9 +1,0 @@
-using DigitalGuardBook.Data.Entities;
-
-namespace DigitalGuardBook.Repositories;
-
-public interface IStationRepository
-{
-    Task<IList<Station>> AllStationsAsync();
-    Task<Station> GetStationAsync();
-}
